ok, forgot to commit merge - so now like this.
The main thing this code does is to add a bit of.... well, evil trickery.

We really do want to validate certificates without cached information. I have no idea how to tell NSS to un-cache stuff. Fix: reinit it.

God, I hope we will be able to get rid of this stupid code.